requirements.txt
openai>=1.14
jinja2
requests
gitpython
